+++ mcpower_gui/__init__.py
@@ -5,4 +5,4 @@ try:
 
     __version__ = _get_version("mcpower-gui")
 except Exception:
-    __version__ = "0.3.0"
+    __version__ = "0.3.1"